+++ example/data-persistence-tf/outputs.tf
@@ -3,22 +3,6 @@ output "dynamo_tables" {
   value = module.data_persistence.dynamo_tables
 }
 
-output "elasticsearch_domain_arn" {
-  value = module.data_persistence.elasticsearch_domain_arn
-}
-
-output "elasticsearch_hostname" {
-  value = module.data_persistence.elasticsearch_hostname
-}
-
-output "elasticsearch_security_group_id" {
-  value = module.data_persistence.elasticsearch_security_group_id
-}
-
-output "elasticsearch_alarms" {
-  value = module.data_persistence.elasticsearch_alarms
-}
-
 output "database_credentials_secret_arn" {
   value = module.provision_database.database_credentials_secret_arn
 }